Move ApplicationList component to separate file and export from index
