+++ app/src/features/ApplicationList/index.ts
@@ -1,0 +1,1 @@
+export {ApplicationList} from './ApplicationList';
